resolved PATH error
This pull resolves an error caused by improper handling of the script's
current location during configuration file inclusion by alternate users.
This caused errors when executing the script either from the command
line or from crontab in certain situations. This pull resolves those,
allowing the script to identify the location of the configuration file
so long as it, or a symlink to it, remains in the same directory as the
script itself, no matter which user executes the script or where it is
stored.